pyproject.toml
[project]
name = "pyauto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]",
    "pyautogui",
    "opencv-python",
    "pillow",
]
